+++ acb-sdk/pluginset/mychain0.20/offchain-plugin/src/test/java/com/alipay/antchain/bridge/plugins/mychain020/common/BizContractTypeEnum.java
@@ -1,0 +1,5 @@
+package com.alipay.antchain.bridge.plugins.mychain020.common;
+
+public enum BizContractTypeEnum {
+    SOLIDITY, WASM, FABRIC, TEEWASM
+}